getCourseId updated

+++ src/course.java
@@ -5,6 +5,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import javax.swing.JComboBox;
 import javax.swing.JOptionPane;
 import javax.swing.JTable;
 import javax.swing.table.DefaultTableModel;
@@ -56,7 +57,10 @@ public class course {
             }
     }
      if(operation == 'd'){//d for delete
-            
+            int YesOrNo = JOptionPane.showConfirmDialog(null, "The Scores Will Be Also Deleted","Delete Score",JOptionPane.OK_CANCEL_OPTION,0);
+         
+        if(YesOrNo == JOptionPane.OK_OPTION)
+        {
             try { 
                 ps = con.prepareStatement("DELETE FROM `course` WHERE `id` = ?" );
                 ps.setInt(1,id);
@@ -64,13 +68,10 @@ public class course {
                 if(ps.executeUpdate()> 0){
                     JOptionPane.showMessageDialog(null,"Course Deleted");
                 }
-                
-                
-                
-                
             } catch (SQLException ex) {
                 Logger.getLogger(student.class.getName()).log(Level.SEVERE, null, ex);
             }
+        }
     }
     }
      public boolean isCourseExist(String CourseName)
@@ -97,7 +98,7 @@ public class course {
          return isExist;
      }
      
-      public void fillCourseJtable(JTable table)
+    public void fillCourseJtable(JTable table)
     {
         Connection con = MyConnection.getConnection();
         PreparedStatement ps;
@@ -117,13 +118,52 @@ public class course {
                 
             }
             
-            
-            
         } catch (SQLException ex) {
             Logger.getLogger(student.class.getName()).log(Level.SEVERE, null, ex);
         }
         
     }
+    public int getCourseId(String courseLabel){
+
+     int courseId = 0;
      
+      Connection con = MyConnection.getConnection();
+        PreparedStatement ps;
+        try {
+            ps = con.prepareStatement("SELECT * FROM `course` WHERE `label` = ?");
+            ps.setString(1,courseLabel);
+            
+            ResultSet rs = ps.executeQuery();
+         
+            if(rs.next()){
+                courseId = rs.getInt("Id");   
+            }
+        } catch (SQLException ex) {
+            Logger.getLogger(student.class.getName()).log(Level.SEVERE, null, ex);
+        }
+     
+     
+     
+     return courseId;
     
+      }
+        
+    public void fillCourseCombo(JComboBox combo)
+    {
+        Connection con = MyConnection.getConnection();
+        PreparedStatement ps;
+        try {
+            ps = con.prepareStatement("SELECT  *FROM `course` ");
+            ResultSet rs = ps.executeQuery();
+           
+            while(rs.next()){
+              combo.addItem(rs.getString(2));  
+            }
+            
+        } catch (SQLException ex) {
+            Logger.getLogger(student.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+      
+
 }
